pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Taushicki <ipanteleev193@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.3"
beautifulsoup4 = "^4.12.3"
tortoise-orm = "^0.20.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
environs = "^10.3.0"
asyncpg = "^0.29.0"
python-binance = "^1.0.19"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
